+++ mastermind/src/gamemanagement/AutoCodeCreater.py
