+++ mastermind/src/gamemanagement/IScene.py
