+++ mastermind/src/gamemanagement/SceneManager.py
@@ -3,7 +3,7 @@ Created on 29.07.2017
 
 @author: Mirco
 '''
-from view.TitleScene import TitleScene
+from metagame.TitleScene import TitleScene
 
 class SceneManager(object):
