+++ mastermind/src/ingame/CrypterScene.py
@@ -3,11 +3,11 @@ Created on 30.07.2017
 
 @author: Mirco
 '''
-from view.IScene import IScene
+from gamemanagement.IScene import IScene
 from settings.ViewSettings import COLORS_TO_RGB, BLOCK_SIZE
-from view.Stone import Stone
 from settings import Settings
 import pygame
+from ingame.Stone import Stone
 
 class CrypterScene(IScene):
     '''
