+++ mastermind/src/ingame/EncrypterScene.py
@@ -6,11 +6,11 @@ Created on 29.07.2017
 import pygame
 from settings import Settings
 from settings.ViewSettings import COLORS_TO_RGB
-from view.Stone import Stone
 from settings.ViewSettings import BLOCK_SIZE
-from view.DecrypterScene import DecrypterScene
-from view.CrypterScene import CrypterScene
 from settings import ViewSettings
+from ingame.CrypterScene import CrypterScene
+from ingame.DecrypterScene import DecrypterScene
+from ingame.Stone import Stone
 
 
 class EncrypterScene(CrypterScene):
