+++ mastermind/src/ingame/Stone.py
